Api

PortfolioOpt

PortfolioOpt.TestUtils

PortfolioOpt.MeanVariance

PortfolioOpt.RobustBenTal

PortfolioOpt.RobustBertsimas

PortfolioOpt.RobustBetina

PortfolioOpt.RobustDelague

PortfolioOpt.SampleBased

PortfolioOpt.base_model

PortfolioOpt.compute_solution

PortfolioOpt.conditional_expectation!

PortfolioOpt.create_sum_invested_variable

PortfolioOpt.equal_weights

PortfolioOpt.max_sharpe

PortfolioOpt.mean_variance_noRf_analytical

PortfolioOpt.mixed_signals_predict_return

PortfolioOpt.po_max_return_limit_variance

PortfolioOpt.po_max_return_limit_variance

PortfolioOpt.po_max_utility_return

PortfolioOpt.po_min_variance_limit_return

PortfolioOpt.po_min_variance_limit_return

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_variance!

PortfolioOpt.predicted_portfolio_return!

PortfolioOpt.predicted_portfolio_variance!

PortfolioOpt.readjust_volumes!

PortfolioOpt.MeanVariance — Type

MeanVariance <: AbstractMeanVariance

Mean Variance Formulation (Markowitz).

PS.: Equivalent to a single point uncertainty set.

Atributes:

predicted_mean::Array{Float64,1}: Predicted mean of returns.

predicted_covariance::Array{Float64,2}: Predicted covariance of returns.

PortfolioOpt.RobustBenTal — Type

BenTal <: AbstractMeanVariance

BenTal's uncertainty set:

\[\left\{ \mu \; \middle| \begin{array}{ll} s.t. \quad \sqrt{(\hat{r} - \mu) ' \Sigma^{-1} (\hat{r} - \mu)} \leq \delta \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

uncertainty_delta::Float64 (latex notation $\delta$): Uncertainty around mean.

predicted_covariance::Array{Float64,2} (latex notation $\Sigma$): Predicted covariance of returns.

PortfolioOpt.RobustBertsimas — Type

RobustBertsimas <: AbstractMeanVariance

Bertsimas's uncertainty set:

\[\left\{ \mu \; \middle| \begin{array}{ll} s.t. \quad \mu_i \leq \hat{r}_i + z_i \Delta_i \quad \forall i = 1:\mathcal{N} \\ \quad \quad \mu_i \geq \hat{r}_i - z_i \Delta_i \quad \forall i = 1:\mathcal{N} \\ \quad \quad z_i \geq 0 \quad \forall i = 1:\mathcal{N} \\ \quad \quad z_i \leq 1 \quad \forall i = 1:\mathcal{N} \\ \quad \quad \sum_{i}^{\mathcal{N}} z_i \leq \Gamma \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

uncertainty_delta::Array{Float64,1} (latex notation $\Delta$): Uncertainty around mean.

bertsimas_budjet::Array{Float64,1} (latex notation $\Gamma$): Number of assets in worst case.

predicted_covariance::Array{Float64,2}: Predicted covariance of returns.

PortfolioOpt.RobustBetina — Type

Betina Robust Sampled based Formulation

PortfolioOpt.RobustDelague — Type

RobustDelague <: AbstractMeanVariance

Delague's Ambiguity set:

\[\left\{ r \; \middle| \begin{array}{ll} s.t. \quad (\mathbb{E} [r] - \hat{r}) ' \Sigma^{-1} (\mathbb{E} [r] - \hat{r}) \leq \gamma_1 \\ \quad \quad \mathbb{E} [ (r - \hat{r}) ' (r - \hat{r}) ] \leq \gamma_2 \Sigma \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

predicted_covariance::Array{Float64,2} (latex notation $\Sigma$): Predicted covariance of returns.

γ1::Float64: Mean uncertainty parameter (has to be greater than 0).

γ2::Float64: Covariance uncertainty parameter (has to be greater than 1).

utility_coeficients::Array{Float64,1}: Piece-wise utility coeficients (default 1).

utility_intercepts::Array{Float64,1}: Piece-wise utility intercepts (default 0).

PortfolioOpt.SampleBased — Type

Betina Robust Sampled based Formulation

PortfolioOpt.base_model — Method

base_model(numA) -> model

Creates a JuMP model with generic PO variable and constraints: - Investment vector of variables w (portfolio weights if current_wealth = 1). - Invested monney should be lower current wealth.

Returns the model (which has a reference to the decision variable named w).

Arguments:

numA::Integer: number of assets in portfolio.

Optional Keywork Arguments:

current_wealth::Real = 1.0: Current available wealth to be invested.

allow_borrow::Bool=false: Short variables work as borrowing money.

PortfolioOpt.compute_solution — Method

compute_solution(model, solver) -> Array{Float64,1}

Solves optimization model using provides solver and returns solutions ajusted to current wealth.

Arguments:

model::JuMP.Model: JuMP portfolio optimization model.

solver: Appropriate solver for the provided model.

Optional Keywork Arguments:

current_wealth::Real = 1.0: Current available wealth to be invested.

decision_variable = model[:w]: Portfolio optimization investment variable reference.

PortfolioOpt.conditional_expectation! — Method

conditionalexpectation = -cvar = -expectedshortfall

PortfolioOpt.create_sum_invested_variable — Method

create_sum_invested_variable(model::JuMP.Model, w)

Creates decision variable representing the invested money in variable-return assets: $||w||$.

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: portfolio optimization investment variable ("weights").

PortfolioOpt.equal_weights — Method

Equal weights.

PortfolioOpt.max_sharpe — Method

Maximize sharp coeficient alocation.

PortfolioOpt.mean_variance_noRf_analytical — Method

Mean-Variance Portfolio Alocation With no risk free asset. Analytical solution.

PortfolioOpt.mixed_signals_predict_return — Function

Mixed signals predictor proposed by Betina and al. Univariate

PortfolioOpt.po_max_return_limit_variance — Method

po_max_return_limit_variance(formulation::AbstractPortfolioFormulation, V_0)

Mean-Variance Portfolio Alocation (with a risk free asset). Posed as a quadratic convex problem. Maximizes worst case portfolio return ($R$) and limit worst case portfolio variance ($V$) in the uncertainty set ($\Omega$) to a maximal risk parameter ($V_0$) normalized by current wealth ($W_0$).

\[\begin{aligned} \max_{w} \quad & R \\ s.t. \quad & R = (\min r'w \; | \; r \in \Omega) \\ & V \leq V_0 * W_0\\ & V = (\max w ' \Sigma w \; | \; \Sigma \in \Omega) \\ & w \in \mathcal{X} \\ \end{aligned}\]

Where $\mathcal{X}$ represents the additional constraints defined in the model by the user (e.g. a limit on maximum invested money).

Arguments:

formulation::AbstractPortfolioFormulation: Struct containing atributes of formulation.

max_risk::Real: Maximal normalized variance accepted.

Optional Keywork Arguments:

rf::Real = 0.0: Risk-free return of money not invested.

current_wealth::Real = 1.0: Current available wealth to be invested.

model::JuMP.Model = base_model(...): JuMP upper level portfolio optimization model. Defaults to a generic market specification.

w: Portfolio optimization investment variable reference ("weights").

portfolio_return::Function = portfolio_return!: Function that calculates $R$.

portfolio_variance::Function = portfolio_variance!: Function that calculates $V$.

PortfolioOpt.po_max_utility_return — Method

po_max_utility_return(formulation::RobustDelague)

Maximize expected utility of portfolio return under distribution uncertainty defined by Delague's ambiguity set (RobustDelague).

PortfolioOpt.po_min_variance_limit_return — Method

po_min_variance_limit_return(formulation::AbstractPortfolioFormulation, R)

Mean-Variance Portfolio Alocation (with a risk free asset). Posed as a quadratic convex problem. Minimizes worst case portfolio variance ($V$) and limit worst case portfolio return ($R$) in the uncertainty set ($\Omega$) to a minimal return parameter ($R_0$) normalized by current wealth ($W_0$).

\[\begin{aligned} \min_{w} \quad & V \\ s.t. \quad & R = (\min r'w \; | \; r \in \Omega) \\ & V = (\max w ' \Sigma w \; | \; \Sigma \in \Omega) \\ & R \geq R * W_0 \\ & w \in \mathcal{X} \\ \end{aligned}\]

Where $\mathcal{X}$ represents the additional constraints defined in the model by the user (e.g. a limit on maximum invested money).

Arguments:

formulation::AbstractPortfolioFormulation: Struct containing atributes of formulation.

minimal_return::Real: Minimal normalized return accepted.

Optional Keywork Arguments:

rf::Real = 0.0: Risk-free return of money not invested.

current_wealth::Real = 1.0: Current available wealth to be invested.

model::JuMP.Model = base_model(...): JuMP upper level portfolio optimization model. Defaults to a generic market specification.

w: Portfolio optimization investment variable reference ("weights").

portfolio_return::Function = portfolio_return!: Function that calculates $R$.

portfolio_variance::Function = portfolio_variance!: Function that calculates $V$.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::AbstractMeanVariance; kwargs...)

Return worst case portfolio return of the Uncertainty set defined by the formulation.

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: Portfolio optimization investment variable ("weights").

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBenTal)

Returns worst case return (WCR) in BenTal's uncertainty set (RobustBenTal).

WCR is defined by the following primal problem:

\[\begin{aligned} \min_{\mu, z} \quad & \mu ' w \\ s.t. \quad & ||Σ^{-\frac{1}{2}} (\mu - \hat{r}) || \leq \delta \quad &: \theta \\ \end{aligned}\]

Which is equivalent to the following dual problem:

\[\begin{aligned} \max_{\theta} \quad & w ' \hat{r} - \theta \delta \\ s.t. \quad & ||Σ^{\frac{1}{2}} w || \leq \theta \\ \end{aligned}\]

To avoid solving an optimization problem we enforece the dual constraints in the upper level problem and return the objective expression (a lower bound of the optimum).

Arguments: - model::JuMP.Model: JuMP upper level portfolio optimization model. - w: portfolio optimization investment variable ("weights"). - formulation::RobustBenTal: Struct containing atributes of BenTal's uncertainty set.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBertsimas)

Returns worst case return (WCR) in Bertsimas's uncertainty set (RobustBertsimas).

WCR is defined by the following primal problem:

\[\begin{aligned} \min_{\mu, z} \quad & \mu ' w \\ s.t. \quad & \mu_i \leq \hat{r}_i + z_i \Delta_i \quad \forall i = 1:\mathcal{N} \quad &: \pi^-_i \\ & \mu_i \geq \hat{r}_i - z_i \Delta_i \quad \forall i = 1:\mathcal{N} \quad &: \pi^+_i \\ & z_i \geq 0 \quad \forall i = 1:\mathcal{N} \\ & z_i \leq 1 \quad \forall i = 1:\mathcal{N} \quad &: \theta_i \\ & \sum_{i}^{\mathcal{N}} z_i \leq \Gamma \quad : \lambda \\ \end{aligned}\]

Which is equivalent to the following dual problem:

\[\begin{aligned} \max_{\lambda, \pi^-, \pi^+, \theta} \quad \sum_{i}^{\mathcal{N}} (\hat{r}_i (\pi^+_i - \pi^-_i) - \theta_i ) - \Gamma \lambda\\ s.t. \quad & w_i = \pi^+_i - \pi^-_i \quad \forall i = 1:\mathcal{N} \\ & \Delta_i (\pi^+_i + \pi^-_i) - \theta_i \leq \lambda \quad \forall i = 1:\mathcal{N} \\ & \lambda \geq 0 , \; \pi^- \geq 0 , \; \pi^+ \geq 0 , \; \theta \geq 0 \\ \end{aligned}\]

To avoid solving an optimization problem we enforece the dual constraints in the upper level problem and return the objective expression (a lower bound of the optimum).

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: portfolio optimization investment variable ("weights").

formulation::RobustBertsimas: Struct containing atributes of Bertsimas's uncertainty set.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBetina)

Returns worst case return in Betina's uncertainty set, defined by the following dual problem:

\[\max_{\theta} \quad \theta \\ s.t. \quad \theta \leq r_s ' w \quad \forall s = 1:\mathcal{S} \\\]

PortfolioOpt.portfolio_variance! — Method

portfolio_variance!(model::JuMP.Model, w, formulation::AbstractMeanVariance; kwargs...)

Return worst case portfolio variance of the Uncertainty set defined by the formulation.

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: Portfolio optimization investment variable ("weights").

PortfolioOpt.predicted_portfolio_return! — Method

predicted_portfolio_return!(model::JuMP.Model, w, formulation::AbstractMeanVariance)

Return predicted portfolio return used in the formulation.

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: Portfolio optimization investment variable ("weights").

PortfolioOpt.predicted_portfolio_variance! — Method

predicted_portfolio_variance!(model::JuMP.Model, w, formulation::AbstractMeanVariance)

Return predicted portfolio variance used in the formulation:

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: Portfolio optimization investment variable ("weights").

PortfolioOpt.readjust_volumes! — Method

readjust_volumes(decision_values::Array{Float64,1}; current_wealth::Real = 1.0)

Ajust volumes to be feasible under current wealth.

Arguments:

decision_values::Array{Float64,1}: Decided values to be invested.

Optional Keywork Arguments:

current_wealth::Real = 1.0: Current available wealth to be invested.

TestUtils

PortfolioOpt.TestUtils — Module

TestUtils

Some commom test utilities for portfolio optimization formulations.

Mainly: - get_test_data that returns a TimeArray of Prices for 6 assets. - backtest_po that provides a basic backtest using provided strategy and returns data.

PortfolioOpt.TestUtils.backtest_po — Method

backtest_po(strategy_logic::Function, returns_series::TimeArray) -> TimeArray, Array{Float64}

Simple backtest functionality for strategies that return an array of invested money per asset.

API:

wealth_strategy, returns_strategy = backtest_po(returns_series; start_date=start_date) 
    do past_returns, current_wealth, risk_free_return
    # ... Strategy definition ...
    return investment_decision
end

Arguments:

strategy_logic::Function: Function that represents investment strategy.

returns_series::TimeArray{Float64,2,Dates.Date,Array{Float64,2}}: Asset return series.

Optional Keywork Arguments:

start_date::Date=timestamp(returns_series)[1]: Start date of the backtest.

end_date::Date=timestamp(returns_series)[end]: End date of the backtest.

initial_wealth::Real=1.0: Initial available wealth to be invested.

risk_free_returns::TimeArray=TimeArray(timestamp(returns_series), fill(0.0, size(returns_series, 1))): Risk-free returns of money not invested per period.

PortfolioOpt.TestUtils.get_test_data — Method

Get test data (Prices) from MarketData.

PortfolioOpt.TestUtils.mean_variance — Method

Mean and Variance of returns