Api

PortfolioOpt

PortfolioOpt.TestUtils

PortfolioOpt.MeanVariance

PortfolioOpt.RobustBenTal

PortfolioOpt.RobustBertsimas

PortfolioOpt.RobustBetina

PortfolioOpt.RobustDelague

PortfolioOpt.SampleBased

PortfolioOpt.conditional_expectation!

PortfolioOpt.equal_weights

PortfolioOpt.max_sharpe

PortfolioOpt.mean_variance_noRf_analytical

PortfolioOpt.mixed_signals_predict_return

PortfolioOpt.po_max_return_limit_variance!

PortfolioOpt.po_max_return_limit_variance!

PortfolioOpt.po_max_utility_return!

PortfolioOpt.po_min_variance_limit_return!

PortfolioOpt.po_min_variance_limit_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_return!

PortfolioOpt.portfolio_variance!

PortfolioOpt.predicted_portfolio_return!

PortfolioOpt.predicted_portfolio_variance!

PortfolioOpt.MeanVariance — Type

MeanVariance <: AbstractMeanVariance

Mean Variance Formulation (Markowitz).

PS.: Equivalent to a single point uncertainty set.

Atributes:

predicted_mean::Array{Float64,1}: Predicted mean of returns.

predicted_covariance::Array{Float64,2}: Predicted covariance of returns.

PortfolioOpt.RobustBenTal — Type

BenTal <: AbstractMeanVariance

BenTal's uncertainty set:

\[\left\{ \mu \; \middle| \begin{array}{ll} s.t. \quad \sqrt{(\hat{r} - \mu) ' \Sigma^{-1} (\hat{r} - \mu)} \leq \delta \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

uncertainty_delta::Float64 (latex notation $\delta$): Uncertainty around mean.

predicted_covariance::Array{Float64,2} (latex notation $\Sigma$): Predicted covariance of returns.

PortfolioOpt.RobustBertsimas — Type

RobustBertsimas <: AbstractMeanVariance

Bertsimas's uncertainty set:

\[\left\{ \mu \; \middle| \begin{array}{ll} s.t. \quad \mu_i \leq \hat{r}_i + z_i \Delta_i \quad \forall i = 1:\mathcal{N} \\ \quad \quad \mu_i \geq \hat{r}_i - z_i \Delta_i \quad \forall i = 1:\mathcal{N} \\ \quad \quad z_i \geq 0 \quad \forall i = 1:\mathcal{N} \\ \quad \quad z_i \leq 1 \quad \forall i = 1:\mathcal{N} \\ \quad \quad \sum_{i}^{\mathcal{N}} z_i \leq \Gamma \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

uncertainty_delta::Array{Float64,1} (latex notation $\Delta$): Uncertainty around mean.

bertsimas_budjet::Array{Float64,1} (latex notation $\Gamma$): Number of assets in worst case.

predicted_covariance::Array{Float64,2}: Predicted covariance of returns.

PortfolioOpt.RobustBetina — Type

Betina Robust Sampled based Formulation

PortfolioOpt.RobustDelague — Type

RobustDelague <: AbstractMeanVariance

Delague's Ambiguity set:

\[\left\{ r \; \middle| \begin{array}{ll} s.t. \quad (\mathbb{E} [r] - \hat{r}) ' \Sigma^{-1} (\mathbb{E} [r] - \hat{r}) \leq \gamma_1 \\ \quad \quad \mathbb{E} [ (r - \hat{r}) ' (r - \hat{r}) ] \leq \gamma_2 \Sigma \\ \end{array} \right\} \\\]

Atributes:

predicted_mean::Array{Float64,1} (latex notation $\hat{r}$): Predicted mean of returns.

predicted_covariance::Array{Float64,2} (latex notation $\Sigma$): Predicted covariance of returns.

γ1::Float64: Mean uncertainty parameter (has to be greater than 0).

γ2::Float64: Covariance uncertainty parameter (has to be greater than 1).

utility_coeficients::Array{Float64,1}: Piece-wise utility coeficients (default 1).

utility_intercepts::Array{Float64,1}: Piece-wise utility intercepts (default 0).

PortfolioOpt.SampleBased — Type

Betina Robust Sampled based Formulation

PortfolioOpt.conditional_expectation! — Method

conditionalexpectation = -cvar = -expectedshortfall

PortfolioOpt.equal_weights — Method

Equal weights.

PortfolioOpt.max_sharpe — Method

Maximize sharp coeficient alocation.

PortfolioOpt.mean_variance_noRf_analytical — Method

Mean-Variance Portfolio Alocation With no risk free asset. Analytical solution.

PortfolioOpt.mixed_signals_predict_return — Function

Mixed signals predictor proposed by Betina and al. Univariate

PortfolioOpt.po_max_return_limit_variance! — Method

po_max_return_limit_variance!(model::JuMP.Model, w, formulation::AbstractPortfolioFormulation, max_risk)

Mean-Variance Portfolio Alocation (with a risk free asset). Posed as a quadratic convex problem. Maximizes worst case portfolio return (WCR) and limit worst case portfolio variance (WCV) in uncertainty set.

\[\max_{w} \quad WCR \\ s.t. \quad WCR = (\min r'w \; | \; r \in UncertaintySet) \\ \quad \quad WCV \leq MaxRisk * CurrentWealth\\ \quad \quad WCV = (\max w ' \Sigma w \; | \; \Sigma \in UncertaintySet) \\ \quad \quad w \in \mathcal{X} \\\]

Where $\mathcal{X}$ represents the additional constraints defined in the model by the user (like maximum invested money).

PortfolioOpt.po_max_utility_return! — Method

po_max_utility_return!(model::JuMP.Model, w, formulation::RobustDelague)

Maximize expected utility of portfolio return under distribution uncertainty defined by Delague's ambiguity set (RobustDelague).

PortfolioOpt.po_min_variance_limit_return! — Method

po_min_variance_limit_return!(model::JuMP.Model, w, formulation::AbstractPortfolioFormulation, R)

Mean-Variance Portfolio Alocation (with a risk free asset). Posed as a quadratic convex problem. Minimizes worst case portfolio variance (WCV) and limit worst case portfolio return (WCR) in uncertainty set.

\[\min_{w} \quad WCV \\ s.t. \quad WCR = (\min r'w \; | \; r \in UncertaintySet) \\ \quad \quad WCV = (\max w ' \Sigma w \; | \; \Sigma \in UncertaintySet) \\ \quad \quad WCR \geq R * CurrentWealth \\ \quad \quad w \in \mathcal{X} \\\]

Where $\mathcal{X}$ represents the additional constraints defined in the model by the user (like maximum invested money).

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::AbstractMeanVariance; kwargs...)

Return worst case return of the Uncertainty set defined by the formulation.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBenTal)

Returns worst case return (WCR) in BenTal's uncertainty set (RobustBenTal).

WCR is defined by the following primal problem:

\[\min_{\mu, z} \mu ' w \\ s.t. \quad ||Σ^{-\frac{1}{2}} (\mu - \hat{r}) || \leq \delta \quad : \theta \\\]

Which is equivalent to the following dual problem:

\[\max_{\theta} \quad w ' \hat{r} - \theta ' \delta \\ s.t. \quad ||Σ^{\frac{1}{2}} w || \leq \theta \\\]

To avoid solving an optimization problem we enforece the dual constraints in the upper level problem and return the objective expression (a lower bound of the optimum).

Arguments: - model::JuMP.Model: JuMP upper level portfolio optimization model. - w: portfolio optimization investment variable ("weights"). - formulation::RobustBenTal: Struct containing atributes of BenTal's uncertainty set.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBertsimas)

Returns worst case return (WCR) in Bertsimas's uncertainty set (RobustBertsimas).

WCR is defined by the following primal problem:

\[\min_{\mu, z} \mu ' w \\ s.t. \quad \mu_i \leq \hat{r}_i + z_i \Delta_i \quad \forall i = 1:\mathcal{N} \quad : \pi^-_i \\ \quad \quad \mu_i \geq \hat{r}_i - z_i \Delta_i \quad \forall i = 1:\mathcal{N} \quad : \pi^+_i \\ \quad \quad z_i \geq 0 \quad \forall i = 1:\mathcal{N} \\ \quad \quad z_i \leq 1 \quad \forall i = 1:\mathcal{N} \quad : \theta_i \\ \quad \quad \sum_{i}^{\mathcal{N}} z_i \leq \Gamma \quad : \lambda \\\]

Which is equivalent to the following dual problem:

\[\max_{\lambda, \pi^-, \pi^+, \theta} \quad \sum_{i}^{\mathcal{N}} (\hat{r}_i (\pi^+_i \pi^-_i) - \theta_i ) - \Gamma \lambda\\ s.t. \quad w_i = \pi^+_i - \pi^-_i \quad \forall i = 1:\mathcal{N} \\ \quad \quad \Delta_i (\pi^+_i + \pi^-_i) - \theta_i \leq \lambda \quad \forall i = 1:\mathcal{N} \\ \quad \lambda \geq 0 , \; \pi^- \geq 0 , \; \pi^+ \geq 0 , \; \theta \geq 0 \\\]

To avoid solving an optimization problem we enforece the dual constraints in the upper level problem and return the objective expression (a lower bound of the optimum).

Arguments:

model::JuMP.Model: JuMP upper level portfolio optimization model.

w: portfolio optimization investment variable ("weights").

formulation::RobustBertsimas: Struct containing atributes of Bertsimas's uncertainty set.

PortfolioOpt.portfolio_return! — Method

portfolio_return!(model::JuMP.Model, w, formulation::RobustBetina)

Returns worst case return in Betina's uncertainty set, defined by the following dual problem:

\[\max_{\theta} \quad \theta \\ s.t. \quad \theta \leq r_s ' w \quad \forall s = 1:\mathcal{S} \\\]

PortfolioOpt.portfolio_variance! — Method

portfolio_variance!(model::JuMP.Model, w, formulation::AbstractMeanVariance; kwargs...)

Return worst case variance of the Uncertainty set defined by the formulation.

PortfolioOpt.predicted_portfolio_return! — Method

predicted_portfolio_return!(model::JuMP.Model, w, formulation::AbstractMeanVariance)

Return predicted return used in the formulation.

PortfolioOpt.predicted_portfolio_variance! — Method

predicted_portfolio_variance!(model::JuMP.Model, w, formulation::AbstractMeanVariance)

Return predicted variance used in the formulation.

TestUtils

PortfolioOpt.TestUtils — Module

TestUtils

Some commom test utilities for portfolio optimization formulations.

Mainly: - get_test_data that returns a TimeArray of Prices for 6 assets. - backtest_po that provides a basic backtest using provided strategy and returns data.

PortfolioOpt.TestUtils.backtest_po — Method

Simple backtest logic

PortfolioOpt.TestUtils.base_model — Method

base_model(numA) -> model, w

Creates a JuMP model with generic PO variable and constraints: - Investment vector of variables w (portfolio weights if budget is 1). - Invested monney should be lower than a budget.

Returns the model and a reference to the decision variable w.

PortfolioOpt.TestUtils.compute_solution_backtest — Method

Basic solution

PortfolioOpt.TestUtils.get_test_data — Method

Get test data (Prices) from MarketData.

PortfolioOpt.TestUtils.mean_variance — Method

Mean and Variance of returns

PortfolioOpt.TestUtils.readjust_volumes — Method

Ajust volumes to be feasible under current wealth